Renaming test_list_setups header and source files to list_test_setups
This is for the sake of keeping the filename pattern homogeneous
