+++ tests/utils/datastructure/list_create_tests.c
@@ -3,7 +3,7 @@
 #include "utils/datastructure/list.h"
 #include "utils/datastructure/list_private.h"
 #include "utils/datastructure/list_tests.h"
-#include "utils/datastructure/test_list_setups.h"
+#include "utils/datastructure/list_test_setups.h"
 
 DEFINE_STANDALONE_TEST_FUNC(createNode) {
     Node *node = createNode((void*)1UL,
