+++ tests/utils/datastructure/list_getter_tests.c
@@ -3,7 +3,7 @@
 #include "utils/datastructure/list.h"
 #include "utils/datastructure/list_private.h"
 #include "utils/datastructure/list_tests.h"
-#include "utils/datastructure/test_list_setups.h"
+#include "utils/datastructure/list_test_setups.h"
 
 DEFINE_STANDALONE_TEST_FUNC(listGetElementsCountNullPtr) {
     munit_assert_ulong(listGetElementsCount(NULL), ==, 0UL);
