+++ tests/utils/datastructure/list_remove_tests.c
@@ -3,7 +3,7 @@
 #include "utils/datastructure/list.h"
 #include "utils/datastructure/list_private.h"
 #include "utils/datastructure/list_tests.h"
-#include "utils/datastructure/test_list_setups.h"
+#include "utils/datastructure/list_test_setups.h"
 
 DEFINE_STANDALONE_TEST_FUNC(listPopFrontNullPtr) {
     munit_assert_long(listPopFront(NULL), ==, LIST_FAIL);
